requirements.txt
Flask==0.12
Flask-SQLAlchemy==2.1
Flask-DebugToolbar==0.10.1
Markdown==2.6.8
Flask-RESTful==0.3.5
PyYAML==3.12
Cerberus==1.1
marshmallow-sqlalchemy==0.12.1
flask-marshmallow==0.7.0
